requirements.txt
requests==2.28.1
matplotlib==3.5.2
python-dotenv==0.21.0
